Extract intial EDTR registration
