+++ domains/rem/src/hooks/entityActionsMenu/index.ts
@@ -1,0 +1,1 @@
+export { default as useDatesActionsMenuSubscription } from './useDatesActionsMenuSubscription';
