+++ domains/rem/src/hooks/useEditorInit.ts
@@ -1,0 +1,8 @@
+import { useDatesActionsMenuSubscription } from './';
+
+const useEditorInit: VoidFunction = () => {
+	// set menu subscription
+	useDatesActionsMenuSubscription();
+};
+
+export default useEditorInit;
