+++ domains/rem/src/ui/REM.tsx
@@ -1,33 +1,11 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 
-import { useEntityActionsSubscription } from '@eventespresso/registry';
-import { domain, Datetime } from '@eventespresso/edtr-services';
-import { DatetimeProvider } from '../context';
-import RemButton from './RemButton';
-import type { EntityActionsSubscriptionCb } from '@eventespresso/registry';
-
-type DatesSubscriptionCallback = EntityActionsSubscriptionCb<Datetime, 'datetime'>;
+import { useEditorInit } from '../hooks';
 
 const REM: React.FC = () => {
-	const { subscribe } = useEntityActionsSubscription(domain);
+	useEditorInit();
 
-	const datesActionHandler = useCallback<DatesSubscriptionCallback>(({ entity, registry }) => {
-		const { registerElement: registerMenuItem } = registry;
-
-		registerMenuItem(
-			'rem',
-			() => (
-				<DatetimeProvider datetime={entity}>
-					<RemButton />
-				</DatetimeProvider>
-			),
-			11
-		);
-	}, []);
-
-	subscribe(datesActionHandler, { entityType: 'datetime' });
-
-	return null;
+	return <></>;
 };
 
 export default REM;
