+++ domains/rem/src/ui/RemButton.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { useDisclosure } from '@chakra-ui/hooks';
 
-import { withEdtrContext } from '@eventespresso/edtr-services';
 import { IconButton } from '@eventespresso/components';
 import { Rem } from '@eventespresso/icons';
 
 import Modal from '../components/Modal';
+import { useRemInitialization } from '../hooks';
 
 const RemButton: React.FC = () => {
+	useRemInitialization();
 	const { isOpen, onOpen, ...disclosure } = useDisclosure();
 
 	return (
@@ -18,4 +19,4 @@ const RemButton: React.FC = () => {
 	);
 };
 
-export default withEdtrContext(RemButton);
+export default RemButton;
