+++ types/global.ts
@@ -1,4 +1,5 @@
 import type { EventEditorData } from '@eventespresso/edtr-services';
+import type { RemDomData } from '@eventespresso/rem/src/types';
 import type { WpPluginsPageData } from '@eventespresso/wp-plugins-page/src/types';
 import type { EventEspressoDomData } from '@eventespresso/services';
 
@@ -7,6 +8,7 @@ import type { EventEspressoDomData } from '@eventespresso/services';
  */
 export interface EventEspressoData extends EventEspressoDomData {
 	eventEditor?: EventEditorData;
+	rem?: RemDomData;
 	wpPluginsPage?: WpPluginsPageData;
 }
